cleaned up global. added DiscordLog
