+++ TSCommand.js
@@ -1,12 +1,10 @@
 const { Command } = require('discord-akairo');
+const TS = require('./TS.js');
+const DiscordLog = require('./DiscordLog');
 class TSCommand extends Command {
 
     async tsexec(ts,message,args){
 
-    }
-
-    defaultMessage(obj){
-        DEFAULTMESSAGES=Object.assign(DEFAULTMESSAGES,obj)
     }
 
     async canRun(ts,message){
@@ -20,17 +18,16 @@ class TSCommand extends Command {
         }
         let ts;
         try {
-            ts=get_ts(message.guild.id)
+            ts=TS.teams(message.guild.id)
         } catch(error){
             message.reply(error)
             throw error;
         }
         
         if(!await this.canRun(ts,message)){
-            console_error(ts.makeErrorObj(`can't run: ${message.content}`,message))
+            DiscordLog.error(ts.makeErrorObj(`can't run: ${message.content}`,message),ts.client)
             return false;
         }
-        
 
         try{
             args.command=ts.parse_command(message)
